Replaced prevous multiEnvironment implementation as the new one should cater for the original use case as well.
